# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Study-ing API
  version: 0.2.0
  description: |
    Study-ing 프론트엔드와 Spring Boot 백엔드가 공유하는 API 계약입니다.

    구현 원칙:
    - GitLab access token은 브라우저에 반환하지 않습니다.
    - 인증된 API는 HttpOnly 세션 쿠키를 사용합니다.
    - Workspace ID로 연결 프로젝트를 결정하며 요청에서 임의 프로젝트를 받지 않습니다.
    - 날짜는 `YYYY-MM-DD`, 시각은 timezone을 포함한 ISO 8601 문자열입니다.
    - 수정 요청은 `expectedRevision` 또는 `lastCommitId`로 낙관적 잠금을 적용합니다.
  contact:
    name: Study-ing Team
servers:
  - url: http://localhost:8080/api/v1
    description: 로컬 개발 서버
...
paths:
...
  /workspaces/{workspaceId}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    get:
      tags: [Workspaces]
      summary: Workspace 상세
      operationId: getWorkspace
      x-owner: member-1
      x-implementation-status: implemented
      responses:
        "200":
          description: Workspace 상세
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceResponse"
        "403":
          $ref: "#/components/responses/WorkspaceAccessDenied"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [Workspaces]
      summary: Workspace 이름과 설정 수정
      operationId: updateWorkspace
      x-owner: member-1
      x-implementation-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkspaceRequest"
      responses:
        "200":
          description: 수정된 Workspace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceResponse"
        "403":
          $ref: "#/components/responses/WorkspaceAccessDenied"
    delete:
      tags: [Workspaces]
      summary: Workspace 소프트 삭제
      operationId: softDeleteWorkspace
      x-owner: member-1
      x-implementation-status: implemented
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [confirmationName]
              properties:
                confirmationName:
                  type: string
      responses:
        "204":
          description: 소프트 삭제 완료
        "400":
          $ref: "#/components/responses/InvalidRequest"
...
  /workspaces/{workspaceId}/announcements:
    get:
      tags: [Team Feed]
      summary: 현재 게시 중인 공지 목록
      operationId: listWorkspaceAnnouncements
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      responses:
        "200":
          description: 고정 우선 공지 목록
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/WorkspaceAnnouncement" }
    post:
      tags: [Team Feed]
      summary: Workspace 공지 게시
      operationId: createWorkspaceAnnouncement
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnnouncementInput" }
      responses:
        "201":
          description: 생성된 공지
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkspaceAnnouncement" }
  /workspaces/{workspaceId}/announcements/{announcementId}:
    patch:
      tags: [Team Feed]
      summary: 공지 수정
      operationId: updateWorkspaceAnnouncement
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: announcementId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AnnouncementInput" }
      responses:
        "200":
          description: 수정된 공지
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkspaceAnnouncement" }
    delete:
      tags: [Team Feed]
      summary: 공지 보관
      operationId: archiveWorkspaceAnnouncement
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: announcementId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: 공지 보관 완료 }
  /workspaces/{workspaceId}/announcements/{announcementId}/read:
    patch:
      tags: [Team Feed]
      summary: 공지 읽음 처리
      operationId: markWorkspaceAnnouncementRead
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: announcementId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: 읽음 처리 시각 }
  /workspaces/{workspaceId}/messages:
    get:
      tags: [Team Feed]
      summary: Workspace 메시지 조회
      operationId: listWorkspaceMessages
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: date
          in: query
          schema: { type: string, format: date }
        - name: cursor
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: 최신순 커서 페이지
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkspaceMessagePage" }
    post:
      tags: [Team Feed]
      summary: Workspace 메시지 작성
      operationId: createWorkspaceMessage
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MessageInput" }
      responses:
        "201":
          description: 생성된 메시지
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkspaceMessage" }
  /workspaces/{workspaceId}/messages/{messageId}:
    patch:
      tags: [Team Feed]
      summary: 작성자 또는 관리자의 메시지 수정
      operationId: updateWorkspaceMessage
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: messageId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/MessageInput" }
      responses:
        "200":
          description: 수정된 메시지
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkspaceMessage" }
    delete:
      tags: [Team Feed]
      summary: 작성자 또는 관리자의 메시지 소프트 삭제
      operationId: deleteWorkspaceMessage
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: messageId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204": { description: 메시지 삭제 완료 }
  /workspaces/{workspaceId}/documents:
    get:
      tags: [Team Documents]
      summary: 팀 문서 검색과 커서 목록
      operationId: listWorkspaceDocuments
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: query
          in: query
          schema: { type: string, maxLength: 100 }
        - name: cursor
          in: query
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: 수정일 최신순 문서 목록
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkspaceDocumentPage" }
    post:
      tags: [Team Documents]
      summary: 팀 문서 생성
      operationId: createWorkspaceDocument
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DocumentMutation" }
      responses:
        "201":
          description: 생성된 팀 문서
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkspaceDocument" }
  /workspaces/{workspaceId}/documents/{documentId}:
    get:
      tags: [Team Documents]
      summary: 팀 문서 상세 열람
      operationId: getWorkspaceDocument
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: documentId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Markdown 문서 상세
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkspaceDocument" }
    patch:
      tags: [Team Documents]
      summary: 작성자의 팀 문서 수정
      operationId: updateWorkspaceDocument
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: documentId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DocumentMutation" }
      responses:
        "200":
          description: 수정된 문서와 증가한 version
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WorkspaceDocument" }
        "409": { $ref: "#/components/responses/Conflict" }
    delete:
      tags: [Team Documents]
      summary: 작성자의 팀 문서 소프트 삭제
      operationId: deleteWorkspaceDocument
      x-implementation-status: implemented
      parameters:
        - $ref: "#/components/parameters/WorkspaceId"
        - name: documentId
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: expectedVersion
          in: query
          required: true
          schema: { type: integer, minimum: 0 }
      responses:
        "204": { description: 문서 삭제 완료 }
        "409": { $ref: "#/components/responses/Conflict" }
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: sid
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    InvalidRequest:
      description: 요청 형식 또는 비즈니스 입력값 오류
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
          example:
            code: INVALID_REQUEST
            message: 요청값을 확인해 주세요.
            details:
              field: title
            requestId: req_01J
            timestamp: "2026-07-25T12:00:00Z"
...
    WorkspaceAccessDenied:
      description: Workspace 활성 멤버가 아님
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
          example:
            code: WORKSPACE_ACCESS_DENIED
            message: Workspace에 접근할 수 없습니다.
            timestamp: "2026-07-25T12:00:00Z"
...
    NotFound:
      description: Workspace, Session, Item 또는 파일을 찾지 못함
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
          example:
            code: RESOURCE_NOT_FOUND
            message: 요청한 리소스를 찾지 못했습니다.
            timestamp: "2026-07-25T12:00:00Z"
    Conflict:
      description: revision, lastCommitId 또는 중복 리소스 충돌
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiErrorResponse"
          example:
            code: SESSION_REVISION_CONFLICT
            message: 다른 사용자가 먼저 수정했습니다. 최신 내용을 확인해 주세요.
            details:
              expectedRevision: 3
              actualRevision: 4
            timestamp: "2026-07-25T12:00:00Z"
...
  schemas:
    ApiErrorResponse:
      type: object
      required: [code, message, timestamp]
      properties:
        code:
          type: string
          description: 프론트 분기와 테스트에서 사용하는 안정적인 오류 코드
        message:
          type: string
          description: 사용자에게 표시 가능한 한국어 메시지
        details:
          type: object
          additionalProperties: true
          description: 필드 오류나 충돌 최신값
        requestId:
          type: string
          description: 로그 추적 ID
        timestamp:
          type: string
          format: date-time
    AnnouncementInput:
      type: object
      required: [title, body, pinned]
      properties:
        title: { type: string, minLength: 1, maxLength: 120 }
        body: { type: string, minLength: 1, maxLength: 10000 }
        pinned: { type: boolean }
        publishedAt: { type: [string, "null"], format: date-time }
        expiresAt: { type: [string, "null"], format: date-time }
    WorkspaceAnnouncement:
      type: object
      required: [id, authorName, title, body, pinned, publishedAt, updatedAt, canEdit]
      properties:
        id: { type: string, format: uuid }
        authorName: { type: string }
        title: { type: string }
        body: { type: string }
        pinned: { type: boolean }
        publishedAt: { type: string, format: date-time }
        expiresAt: { type: [string, "null"], format: date-time }
        updatedAt: { type: string, format: date-time }
        canEdit: { type: boolean }
    MessageInput:
      type: object
      required: [body]
      properties:
        body: { type: string, minLength: 1, maxLength: 4000 }
        contextDate: { type: [string, "null"], format: date }
    WorkspaceMessage:
      type: object
      required: [id, authorName, contextDate, body, createdAt, updatedAt, edited, canEdit]
      properties:
        id: { type: string, format: uuid }
        authorName: { type: string }
        contextDate: { type: string, format: date }
        body: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        edited: { type: boolean }
        canEdit: { type: boolean }
    WorkspaceMessagePage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/WorkspaceMessage" }
        nextCursor: { type: [string, "null"], format: date-time }
    DocumentMutation:
      type: object
      required: [title, bodyMarkdown]
      properties:
        title: { type: string, minLength: 1, maxLength: 120 }
        bodyMarkdown: { type: string, maxLength: 100000 }
        expectedVersion: { type: [integer, "null"], minimum: 0 }
    WorkspaceDocument:
      type: object
      required: [id, authorName, title, bodyMarkdown, version, createdAt, updatedAt, canEdit]
      properties:
        id: { type: string, format: uuid }
        authorName: { type: string }
        title: { type: string }
        bodyMarkdown: { type: string }
        version: { type: integer, minimum: 0 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        canEdit: { type: boolean }
    WorkspaceDocumentPage:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/WorkspaceDocument" }
        nextCursor: { type: [string, "null"], format: date-time }
...
    WorkspaceStatus:
      type: string
      enum: [ACTIVE, SOFT_DELETED]
    WorkspaceSummary:
      type: object
      required: [id, name, status, gitlabProjectPath, defaultBranch]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          $ref: "#/components/schemas/WorkspaceStatus"
        gitlabProjectPath:
          type: string
        defaultBranch:
          type: string
        repositoryBasePath:
          type: string
        repositorySchemaVersion:
          type: integer
        importMode:
          type: string
          enum: [EMPTY, COMPATIBLE, LEGACY, PARTIALLY_COMPATIBLE]
        deletedAt:
          type: [string, "null"]
          format: date-time
    Workspace:
      allOf:
        - $ref: "#/components/schemas/WorkspaceSummary"
        - type: object
          required: [gitlabProjectId, members, settings]
          properties:
            gitlabProjectId:
              type: integer
              format: int64
            members:
              type: array
              items:
                $ref: "#/components/schemas/WorkspaceMember"
            settings:
              $ref: "#/components/schemas/WorkspaceSettings"
    WorkspaceResponse:
      type: object
      required: [data]
      properties:
        data:
          $ref: "#/components/schemas/Workspace"
    WorkspaceSettings:
      type: object
      properties:
        notifications:
          type: object
          properties:
            deadline:
              type: boolean
            submission:
              type: boolean
            sessionChanged:
              type: boolean
...
    UpdateWorkspaceRequest:
      type: object
      minProperties: 1
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 80
        settings:
          $ref: "#/components/schemas/WorkspaceSettings"
    MemberStatus:
      type: string
      enum: [ACTIVE, LOST_PROJECT_ACCESS, INACTIVE]
    WorkspaceMember:
      type: object
      required:
        [userId, memberId, gitlabUserId, username, displayName, fileName, status]
      properties:
        userId:
          type: string
          format: uuid
        memberId:
          type: string
          pattern: ^[a-z0-9][a-z0-9-]{1,39}$
        gitlabUserId:
          type: integer
          format: int64
        username:
          type: string
        displayName:
          type: string
        fileName:
          type: string
          pattern: ^[a-z0-9][a-z0-9-]{1,39}\.md$
        avatarUrl:
          type: [string, "null"]
          format: uri
        gitlabAccessLevel:
          type: integer
        status:
          $ref: "#/components/schemas/MemberStatus"
...
security:
  - sessionCookie: []